requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
openai>=1.0.0
